# src/main/resources/static/swagger.yaml
openapi: 3.0.3
info:
  title: O-Way API Documentation
  description: Secure Backend API for O-Way Rental System (JWT Based)
  version: 2.0.0

servers:
  - url: http://localhost:8080

paths:

  # =========================
  # AUTH
  # =========================
  /api/auth/register:
    post:
      summary: Register new user (Customer or Driver)
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        "200":
          description: User registered successfully

  /api/auth/login:
    post:
      summary: Login using username and password
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Login successful and JWT returned

  # =========================
  # PROFILE (CUSTOMER)
  # =========================
  /api/customer/profile:
    post:
      summary: Create profile (Logged-in user)
      tags: [Profile]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileRequestDto'
      responses:
        "200":
          description: Profile created successfully

    put:
      summary: Update profile (Logged-in user)
      tags: [Profile]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileRequestDto'
      responses:
        "200":
          description: Profile updated successfully

    get:
      summary: Get my profile
      tags: [Profile]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile fetched successfully

  # =========================
  # VEHICLE (DRIVER)
  # =========================
  /api/driver/vehicle:
    post:
      summary: Create vehicle (Driver only)
      tags: [Vehicle]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleRequestDto'
      responses:
        "200":
          description: Vehicle created successfully

    put:
      summary: Update vehicle (Driver only)
      tags: [Vehicle]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleRequestDto'
      responses:
        "200":
          description: Vehicle updated successfully

    get:
      summary: Get my vehicle
      tags: [Vehicle]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Vehicle fetched successfully

  # =========================
  # RENTAL (CUSTOMER)
  # =========================
  /api/customer/rentals:
    post:
      summary: Create rental (Customer only)
      tags: [Rental]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RentalRequestDto'
      responses:
        "200":
          description: Rental created successfully

  /api/customer/rentals/{rentalId}:
    get:
      summary: Get rental by ID (Owner only)
      tags: [Rental]
      security:
        - bearerAuth: []
      parameters:
        - name: rentalId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rental fetched successfully


# =========================
# COMPONENTS
# =========================
components:

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:

    # ================= COMMON =================
    ApiResponse:
      type: object
      properties:
        success:
          type: integer
        code:
          type: integer
        message:
          type: string
        data:
          type: object

    # ================= AUTH =================
    UserRegisterRequest:
      type: object
      required:
        - name
        - password
        - role
      properties:
        name:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [CUSTOMER, DRIVER]

    LoginRequest:
      type: object
      required:
        - name
        - password
      properties:
        name:
          type: string
        password:
          type: string

    # ================= PROFILE =================
    ProfileRequestDto:
      type: object
      required:
        - fullName
        - email
        - contact
        - dob
        - gender
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        contact:
          type: string
          example: 09123456789
        dob:
          type: string
          format: date
        gender:
          type: string
        location:
          $ref: '#/components/schemas/LocationDto'

    # ================= VEHICLE =================
    VehicleRequestDto:
      type: object
      required:
        - plateNumber
        - contact
        - nrc
        - vehicleStatus
        - driverStatus
      properties:
        plateNumber:
          type: string
        contact:
          type: string
        nrc:
          type: string
        vehicleStatus:
          $ref: '#/components/schemas/VehicleStatus'
        driverStatus:
          $ref: '#/components/schemas/DriverStatus'
        location:
          $ref: '#/components/schemas/LocationDto'
        address:
          $ref: '#/components/schemas/AddressDto'

    # ================= RENTAL =================
    RentalRequestDto:
      type: object
      required:
        - vehicleId
        - rentalTime
        - distance
      properties:
        vehicleId:
          type: integer
        distance:
          type: integer
        rentalTime:
          type: string
          format: date-time
        pickupLocation:
          $ref: '#/components/schemas/LocationDto'
        dropoffLocation:
          $ref: '#/components/schemas/LocationDto'

    # ================= LOCATION =================
    LocationDto:
      type: object
      properties:
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double

    # ================= ADDRESS =================
    AddressDto:
      type: object
      properties:
        city:
          type: string
        township:
          type: string
        road:
          type: string
        street:
          type: string

    # ================= ENUMS =================
    VehicleStatus:
      type: string
      enum:
        - AVAILABLE
        - UNAVAILABLE
        - IN_USE

    DriverStatus:
      type: string
      enum:
        - FREE
        - BUSY
        - OFFLINE

    RentalStatus:
      type: string
      enum:
        - REQUESTED
        - ACCEPTED
        - ONGOING
        - COMPLETED
        - CANCELLED